LayoutTests/inspector/debugger/live-edit-breakpoints-expected.txt

Tests breakpoints are correctly dimmed and restored in JavaScriptSourceFrame during live edit.

Bug 99598
Debugger was enabled.

Running: testEditUndo
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Editing source:
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Undoing source editing:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations

Running: testEditCommit
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Editing source:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Committing edited source:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations

Running: testEditCommitFailEditCommit
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Editing source:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Editing source again so that live edit could succeed:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Committing edited source again:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations

Running: testEditCommitFailUndoCommit
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Editing source:
    OriginalTextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = true)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Committing edited source:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Undoing source editing:
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: debugger
Committing edited source again:
    OriginalTextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.removeBreakpoint(lineNumber = 2)
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
        edit-me-breakpoints.js:2, enabled:true
    Dumping breakpoint locations
        url: edit-me-breakpoints.js, lineNumber: 2, project name: network
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations

Running: testEditAnonymousUndo
Set timer for test function.
Script execution paused.
Script execution resumed.
Showing script source:
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project name: debugger
Editing source:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project name: debugger
Undoing source editing:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project name: debugger
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations

Running: testEditDynamicAnonymousUndo
Set timer for test function.
Script execution paused.
Script execution resumed.
Showing script source:
Setting breakpoint:
    TextEditor.addBreakpoint(lineNumber = 2, disabled = false)
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project name: debugger
Editing source:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project name: debugger
Undoing source editing:
    Dumping breakpoint storage
    Dumping breakpoint locations
        url: , lineNumber: 2, project name: debugger
Finally removing breakpoint:
    TextEditor.removeBreakpoint(lineNumber = 2)
    Dumping breakpoint storage
    Dumping breakpoint locations
Debugger was disabled.
